v002/docker/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# LLM e embeddings
llama-cpp-python==0.2.27
sentence-transformers==2.3.1
gpt4all==2.1.0

# Vector database
faiss-cpu==1.7.4
chromadb==0.4.22

# Document processing
pypdf==4.0.1
python-docx==1.1.0
markdown==3.5.2
beautifulsoup4==4.12.3
lxml==5.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
numpy==1.26.3
pandas==2.1.4
requests==2.31.0
aiofiles==23.2.1
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging
loguru==0.7.2
